packages/openseneca/openseneca-0.1.1.tar.gz/openseneca-0.1.1/tests/test_gpt4.py
import pytest
from unittest.mock import Mock, patch
from openseneca.models.gpt4 import GPT4ChatModel

# TODO: need more details to test the GPT4ChatModel class

def test_gpt4_chat_model_initialization():
    provider_mock = Mock()
    model = GPT4ChatModel(provider=provider_mock)
    assert model.provider == provider_mock

def test_gpt4_chat_model_request():
    provider_mock = Mock()
    provider_mock.request.return_value = "Test response"
    model = GPT4ChatModel(provider=provider_mock)

    response = model.request("Test prompt")
    assert response == "Test response"
    provider_mock.request.assert_called_once()

@patch('openseneca.models.gpt4.GPT4ChatModel.request')
def test_gpt4_chat_model_request_with_instructions(mock_request):
    mock_request.return_value = "Test response"
    model = GPT4ChatModel(provider=Mock())
    response = model.request("Test prompt", instructions="Test instructions")
    assert response == "Test response"
    mock_request.assert_called_once()

@patch('openseneca.models.gpt4.GPT4ChatModel.request')
def test_gpt4_chat_model_request_with_past_messages(mock_request):
    mock_request.return_value = "Test response"
    model = GPT4ChatModel(provider=Mock())
    response = model.request("Test prompt", past_messages=["Past message"])
    assert response == "Test response"
    mock_request.assert_called_once()

@patch('openseneca.models.gpt4.GPT4ChatModel.request')
def test_gpt4_chat_model_request_with_settings(mock_request):
    mock_request.return_value = "Test response"
    model = GPT4ChatModel(provider=Mock())
    response = model.request("Test prompt", settings={"setting": "value"})
    assert response == "Test response"
    mock_request.assert_called_once()